Refactor: combine attacking alone predicate
Crowd of True Believers and Viper, Cruel Conspirator each declared an
identical local AttackingAlonePredicate enum. Extract it into a shared
mage.filter.predicate.permanent.AttackingAlonePredicate and have both
cards use it, removing the duplication.
